harness: absorb subturns into providers
